feat: add custom error handler

+++ src/main/java/dh/backend/proyecto_final_clinica_odontologica/service/impl/OdontologoService.java
@@ -3,6 +3,7 @@ package dh.backend.proyecto_final_clinica_odontologica.service.impl;
 
 
 import dh.backend.proyecto_final_clinica_odontologica.entity.Odontologo;
+import dh.backend.proyecto_final_clinica_odontologica.exception.ResourceNotFoundException;
 import dh.backend.proyecto_final_clinica_odontologica.repository.IOdontologoRepository;
 import dh.backend.proyecto_final_clinica_odontologica.service.IOdontologoService;
 import org.springframework.stereotype.Service;
@@ -36,7 +37,13 @@ public class OdontologoService implements IOdontologoService {
     }
 
     @Override
-    public void eliminarOdontologo(Integer id) {
-        odontologoRepository.deleteById(id);
+    public void eliminarOdontologo(Integer id) throws ResourceNotFoundException {
+        Optional<Odontologo> odontologoOptional = buscarUnOdontologo(id);
+        if(odontologoOptional.isPresent()){
+            odontologoRepository.deleteById(id);
+        }
+        else
+            throw new ResourceNotFoundException("{\"message\": \"odontologo no encontrado\"}");
+
     }
 }
